Move all Ruby extensions into one folder.
