+++ lib/extensions/enumerable.rb
