+++ lib/extensions/object.rb
