+++ lib/extensions/signature.rb
