+++ lib/zajal.rb
@@ -2,9 +2,9 @@ require 'ostruct'
 require 'open-uri'
 
 require_relative "ffipp"
-require_relative "signature"
-require_relative "object"
-require_relative "enumerable"
+require_relative "extensions/signature"
+require_relative "extensions/enumerable"
+require_relative "extensions/object"
+require_relative "extensions/string"
 require_relative "core"
-require_relative "object"
 require_relative "frontends/glfw"
